+++ packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/FlutterAdManagerBannerAd.java
@@ -90,7 +90,7 @@ class FlutterAdManagerBannerAd extends FlutterAd implements FlutterAdLoadedListe
 
     adView.setAdSizes(allSizes);
     adView.setAdListener(new FlutterBannerAdListener(adId, manager, this));
-    adView.loadAd(request.asAdManagerAdRequest());
+    adView.loadAd(request.asAdManagerAdRequest(adUnitId));
   }
 
   public void recordImpression() {