+++ packages/google_mobile_ads/ios/Tests/FLTAppOpenAdTest.m
@@ -33,7 +33,7 @@
   FLTAdRequest *request = OCMClassMock([FLTAdRequest class]);
   OCMStub([request keywords]).andReturn(@[ @"apple" ]);
   GADRequest *gadRequest = OCMClassMock([GADRequest class]);
-  OCMStub([request asGADRequest]).andReturn(gadRequest);
+  OCMStub([request asGADRequest:[OCMArg any]]).andReturn(gadRequest);
 
   [self testLoadShowAppOpenAd:request gadOrGAMRequest:gadRequest];
 }
@@ -42,7 +42,7 @@
   FLTGAMAdRequest *request = OCMClassMock([FLTGAMAdRequest class]);
   OCMStub([request keywords]).andReturn(@[ @"apple" ]);
   GAMRequest *gamRequest = OCMClassMock([GAMRequest class]);
-  OCMStub([request asGAMRequest]).andReturn(gamRequest);
+  OCMStub([request asGAMRequest:[OCMArg any]]).andReturn(gamRequest);
   FLTServerSideVerificationOptions *serverSideVerificationOptions =
       OCMClassMock([FLTServerSideVerificationOptions class]);
   GADServerSideVerificationOptions *gadOptions =
@@ -140,7 +140,7 @@
   FLTAdRequest *request = OCMClassMock([FLTAdRequest class]);
   OCMStub([request keywords]).andReturn(@[ @"apple" ]);
   GADRequest *gadRequest = OCMClassMock([GADRequest class]);
-  OCMStub([request asGADRequest]).andReturn(gadRequest);
+  OCMStub([request asGADRequest:[OCMArg any]]).andReturn(gadRequest);
   [self testFailedToLoad:request];
 }
 
@@ -148,7 +148,7 @@
   FLTGAMAdRequest *request = OCMClassMock([FLTGAMAdRequest class]);
   OCMStub([request keywords]).andReturn(@[ @"apple" ]);
   GAMRequest *gamRequest = OCMClassMock([GAMRequest class]);
-  OCMStub([request asGAMRequest]).andReturn(gamRequest);
+  OCMStub([request asGAMRequest:[OCMArg any]]).andReturn(gamRequest);
   [self testFailedToLoad:request];
 }