+++ packages/google_mobile_ads/test/ad_containers_test.dart
@@ -1328,13 +1328,16 @@ void main() {
       debugDefaultTargetPlatformOverride = TargetPlatform.android;
 
       final AdRequest adRequest = AdRequest(
-          keywords: <String>['1', '2', '3'],
-          contentUrl: 'contentUrl',
-          nonPersonalizedAds: false,
-          neighboringContentUrls: <String>['url1.com', 'url2.com'],
-          httpTimeoutMillis: 12345,
-          location: LocationParams(
-              accuracy: 1.1, longitude: 25, latitude: 38, time: 1));
+        keywords: <String>['1', '2', '3'],
+        contentUrl: 'contentUrl',
+        nonPersonalizedAds: false,
+        neighboringContentUrls: <String>['url1.com', 'url2.com'],
+        httpTimeoutMillis: 12345,
+        location:
+            LocationParams(accuracy: 1.1, longitude: 25, latitude: 38, time: 1),
+        mediationExtrasIdentifier: 'identifier',
+        extras: {'key': 'value'},
+      );
 
       final ByteData byteData = codec.encodeMessage(adRequest)!;
       expect(codec.decodeMessage(byteData), adRequest);
@@ -1344,13 +1347,16 @@ void main() {
       debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
 
       final AdRequest adRequest = AdRequest(
-          keywords: <String>['1', '2', '3'],
-          contentUrl: 'contentUrl',
-          nonPersonalizedAds: false,
-          neighboringContentUrls: <String>['url1.com', 'url2.com'],
-          httpTimeoutMillis: 12345,
-          location: LocationParams(
-              accuracy: 1.1, longitude: 25, latitude: 38, time: 1));
+        keywords: <String>['1', '2', '3'],
+        contentUrl: 'contentUrl',
+        nonPersonalizedAds: false,
+        neighboringContentUrls: <String>['url1.com', 'url2.com'],
+        httpTimeoutMillis: 12345,
+        location:
+            LocationParams(accuracy: 1.1, longitude: 25, latitude: 38, time: 1),
+        mediationExtrasIdentifier: 'identifier',
+        extras: {'key': 'value'},
+      );
 
       final ByteData byteData = codec.encodeMessage(adRequest)!;
       AdRequest decoded = codec.decodeMessage(byteData);
@@ -1359,9 +1365,11 @@ void main() {
       expect(decoded.contentUrl, adRequest.contentUrl);
       expect(decoded.nonPersonalizedAds, adRequest.nonPersonalizedAds);
       expect(decoded.keywords, adRequest.keywords);
+      // Time is not included on iOS.
       expect(decoded.location!.accuracy, 1.1);
       expect(decoded.location!.longitude, 25);
       expect(decoded.location!.latitude, 38);
+      expect(decoded.mediationExtrasIdentifier, 'identifier');
     });
 
     test('encode/decode $LoadAdError', () async {
@@ -1498,7 +1506,7 @@ void main() {
       );
     });
 
-    test('encode/decode AdManagerAdRequest Android', () async {
+    test('encode/decode $AdManagerAdRequest Android', () async {
       debugDefaultTargetPlatformOverride = TargetPlatform.android;
       final AdManagerAdRequest request = AdManagerAdRequest(
         keywords: <String>['who'],
@@ -1513,6 +1521,8 @@ void main() {
         publisherProvidedId: 'test-pub-id',
         location:
             LocationParams(accuracy: 1.1, longitude: 25, latitude: 38, time: 1),
+        mediationExtrasIdentifier: 'identifier',
+        extras: {'key': 'value'},
       );
       final ByteData byteData = codec.encodeMessage(request)!;
 
@@ -1522,7 +1532,7 @@ void main() {
       );
     });
 
-    test('encode/decode AdManagerAdRequest iOS', () async {
+    test('encode/decode $AdManagerAdRequest iOS', () async {
       debugDefaultTargetPlatformOverride = TargetPlatform.iOS;
 
       final AdManagerAdRequest request = AdManagerAdRequest(
@@ -1538,6 +1548,8 @@ void main() {
         publisherProvidedId: 'test-pub-id',
         location:
             LocationParams(accuracy: 1.1, longitude: 25, latitude: 38, time: 1),
+        mediationExtrasIdentifier: 'identifier',
+        extras: {'key': 'value'},
       );
 
       final ByteData byteData = codec.encodeMessage(request)!;
@@ -1550,9 +1562,11 @@ void main() {
       expect(decoded.publisherProvidedId, request.publisherProvidedId);
       expect(decoded.customTargeting, request.customTargeting);
       expect(decoded.customTargetingLists, request.customTargetingLists);
+      // Time is not included on iOS.
       expect(decoded.location!.accuracy, 1.1);
       expect(decoded.location!.longitude, 25);
       expect(decoded.location!.latitude, 38);
+      expect(decoded.mediationExtrasIdentifier, 'identifier');
     });
   });
 }