+++ packages/google_mobile_ads/CHANGELOG.md
@@ -1,3 +1,38 @@
+## 1.0.0
+
+* Mediation is now supported in beta.
+  * There are new APIs to support passing network extras to mediation adapters:
+    * [MediationNetworkExtrasProvider](https://github.com/googleads/googleads-mobile-flutter/blob/master/packages/google_mobile_ads/android/src/main/java/io/flutter/plugins/googlemobileads/MediationNetworkExtrasProvider.java)
+      on Android and [FLTMediationNetworkExtrasProvider](https://github.com/googleads/googleads-mobile-flutter/blob/master/packages/google_mobile_ads/ios/Classes/FLTConstants.h) on iOS
+  * See the mediation example app [README](https://github.com/googleads/googleads-mobile-flutter/blob/master/packages/mediation_example/README.md)
+    for more details on how to use these APIs.
+
+* Fix for Android 12 issue [#330](https://github.com/googleads/googleads-mobile-flutter/issues/330)
+  * This will break compilation on android if you do not already set `compileSdkVersion` to `31`, or override the WorkManager dependency to < 2.7.0:
+      ```
+      dependencies {
+          implementation('androidx.work:work-runtime') {
+              version {
+                  strictly '2.6.0'
+              }
+          }
+      }
+      ```
+* Fixes issue [#404](https://github.com/googleads/googleads-mobile-flutter/issues/404)
+  * Adds a new dart class, `AppStateEventNotifier`. You should subscribe to `AppStateEventNotifier.appStateStream`
+    instead of using `WidgetsBindingObserver` to listen to app foreground/background events.
+  * See the app open [example app](https://github.com/googleads/googleads-mobile-flutter/tree/master/packages/app_open_example) for a reference
+    on how to use the new API.
+    
+* Adds a new parameter `extras` to `AdRequest` and `AdManagerAdRequest`.
+  * This can be used to pass additional signals to the AdMob adapter, such as
+    [CCPA](https://developers.google.com/admob/android/ccpa) signals.
+  * For example, to notify Google that [RDP](https://developers.google.com/admob/android/ccpa#rdp_signal)
+    should be enabled when constructing an ad request:
+    ```dart
+      AdRequest request = AdRequest(extras: {'rdp': '1'});
+    ```
+    
 ## 0.13.6
 
 * Partial fix for [#265](https://github.com/googleads/googleads-mobile-flutter/issues/265).